Allow to hide tab bar when there is a single tab
Attempt to address https://invent.kde.org/plasma-mobile/angelfish/-/issues/88

Disclaimer: I sent the feature request linked above.

This PR adds a setting option to allow the tab bar to be hidden when there is a single tab shown. 

This would allow for a cleaner look when there is just a tab open, which added to the clean look of Angelfish would make it perfect for product screenshots.

I have no previous experience with QML, but I tried to look around similar options to implement it the best way.

Of course suggestions are welcome and I will promptly fix anything wrong if it is within my reach of knowledge.

Lastly, if this not a desired feature to have, there is no hard feelings. It was already fun to look around QML code for the first time.