Port database operations to FutureSQL
This is not yet ready for merge, as FutureSQL itself still needs finalization and KDEReview.

The advantage of using it is mostly to have asynchronous database operations. With large databases, searching the history was already beginning to be blocking the UI.
Additionally, we can use FutureSQLs database migration framework, which makes future migrations a bit less error prone.