Fixed Scrolling in the Tab Drawer and made the New Tab Button Always Switch toa New Tab
The tab drawer on mobile was not able to be scrolled though. Changed ScrollView to Flickable to fix this issues. Also made the BottomDrawer full screen and changed it to not be interactive to prevent other scrolling issues.

If the switch to new tab setting was turned off, the new tab button in the tab drawer would create a new tab, but would send the user to edit the url of the old tab. To fix this, the new tab button now switches the user to the new tab.