Navbar Auto Hide UI Changes
This is an update to the navbar auto hide UI and also prevents the web view from getting blocked by the navbar.

![navbar](/uploads/aec8dbd622569eb6f56a8c6dea861c40/navbar.gif)

The most notable UI difference here is that the navbar will now no longer fully hides, but will instead shrink down. This is so that it can still be tapped on by the user to bring it back to its full size.